Feat: comment likeCnt 추가 및 예외처리

+++ src/main/java/com/example/newsfeed/service/board/BoardService.java
@@ -129,7 +129,7 @@ public class BoardService {
                 comment.getUser().getImgUrl(),
                 comment.getUser().getName(),
                 comment.getContents(),
-                commentLikesCountMap.getOrDefault(comment.getId(), 0L).intValue(),
+                (long) commentLikesCountMap.getOrDefault(comment.getId(), 0L).intValue(),
                 comment.getUpdatedAt()
         ));
